Added the feature of inserting functions.
If we specify a function instead of a string value and the time in minutes, the first time we will execute the function. The next time, we will read from localStorage. After the time has passed, we will execute again the function. This could come in handy if we do **synchronous** ajax queries.